[lab04] run.sh update

+++ lab04/sorts.c
@@ -9,7 +9,7 @@
 #include "algs.h"
 
 int sort_v1(item_t *array, int32_t arr_size, int32_t n_buckets, struct bucket *buckets, unsigned int thread_idx,
-            int8_t num_of_threads) {
+            int32_t num_of_threads) {
 
     double start;
     int width = arr_size / n_buckets;
